requirements.txt
redis==5.0.1
python-jose[cryptography]
fastapi
uvicorn[standard]
httpx
beautifulsoup4
playwright